feat(shared): add global Result<T> type for consistent error handling
Adds shared Result<T> type pattern to frontend/src/shared/types.ts:
- Standardized Result<T> discriminated union type
- isError and isSuccess type guards
- success() and error() utility functions
- CommonErrorCode type for consistent error codes

Updates drill provider to use shared Result type:
- Imports Result from shared/types instead of local definition
- Maintains all functionality and test coverage
- Ensures consistency across the application

This establishes a global error handling pattern that can be used
across all frontend modules for consistent and type-safe error handling.
